+++ src/main/java/qikahome/jsonmore/JsonMore.java
@@ -28,6 +28,7 @@ import qikahome.jsonmore.lib.ingredient.ItemDisplayOverrideIngredient;
 import qikahome.jsonmore.lib.ingredient.KeepInventoryContainerIngredient;
 import qikahome.jsonmore.lib.ingredient.NBTCopyIngredient;
 import qikahome.jsonmore.lib.ingredient.NotIngredient;
+import qikahome.jsonmore.lib.ingredient.RegexIngredient;
 import qikahome.jsonmore.lib.ingredient.RemainderOverrideIngredient;
 import qikahome.jsonmore.lib.ingredient.ToolDamagingIngredient;
 import qikahome.jsonmore.lib.ingredient.TrueIngredient;
@@ -77,6 +78,7 @@ public class JsonMore implements ModInitializer {
 
     static {
         NotIngredient.register();
+        RegexIngredient.register();
         KeepInventoryContainerIngredient.register();
         TrueIngredient.register();
         ToolDamagingIngredient.register();